+++ Src/Apps/Connect.Win.Web/App.cs
@@ -9,18 +9,18 @@ namespace VpnHood.App.Connect.Win.Web;
 
 public class App : Application
 {
-    private static AppOptions CreateAppOptions(AppConfigs appConfigs)
+    private static AppOptions CreateAppOptions()
     {
-        // load app settings and resources
+        var appConfigs = AppConfigs.Load();
         var resources = ConnectAppResources.Resources;
         resources.Strings.AppName = AppConfigs.AppName;
         return new AppOptions(appId: appConfigs.AppId, "VpnHoodConnect", AppConfigs.IsDebugMode) {
-            CustomData = appConfigs.CustomData,
             UiName = "VpnHoodConnect",
+            CustomData = appConfigs.CustomData,
+            WebUiPort = appConfigs.WebUiPort,
             Resources = resources,
             AccessKeys = appConfigs.DefaultAccessKey != null ? [appConfigs.DefaultAccessKey] : [],
             IsAddAccessKeySupported = false,
-            WebUiHostName = "my-vpnhood-connect",
             AllowEndPointTracker = appConfigs.AllowEndPointTracker,
             Ga4MeasurementId = appConfigs.Ga4MeasurementId,
             RemoteSettingsUrl = appConfigs.RemoteSettingsUrl,
@@ -43,10 +43,7 @@ public class App : Application
         base.OnStartup(e);
 
         // load app configs
-        var appConfigs = AppConfigs.Load();
-        VpnHoodAppWpfSpa.Init(() => CreateAppOptions(appConfigs),
-            spaDefaultPort: appConfigs.WebUiPort,
-            args: Environment.GetCommandLineArgs());
+        VpnHoodAppWpfSpa.Init(CreateAppOptions, args: Environment.GetCommandLineArgs());
     }
 
     [STAThread]